requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask
Flask-WTF
Flask-Bootstrap4

# Mongo driver
pymongo

# Environment file handling
python-dotenv